Corrected PK's type

+++ EducationPortal/EducationPortal.BLL/DTO/CourseDTO.cs
@@ -4,13 +4,13 @@
 
     public class CourseDTO
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         public string Name { get; set; }
 
         public string Description { get; set; }
 
-        public int CreatorId { get; set; }
+        public long CreatorId { get; set; }
 
         public IEnumerable<string> SkillNames { get; set; }
     }
