+++ EducationPortal/EducationPortal.DAL/Entities/EF/Course.cs
@@ -4,13 +4,13 @@
 
     public class Course
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         public string Name { get; set; }
 
         public string Description { get; set; }
 
-        public int CreatorId { get; set; }
+        public long CreatorId { get; set; }
 
         public User Creator { get; set; }
